binance-backend/requirements.txt
websocket-client==1.6.4
pytz==2023.3
requests==2.31.0
